% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{AudioTranscriptionsClient}
\alias{AudioTranscriptionsClient}
\title{Audio Transcriptions Client}
\description{
Transcribes audio files to text using OpenAI's Whisper model.
Access via \code{client$audio$transcriptions}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AudioTranscriptionsClient-new}{\code{AudioTranscriptionsClient$new()}}
\item \href{#method-AudioTranscriptionsClient-create}{\code{AudioTranscriptionsClient$create()}}
\item \href{#method-AudioTranscriptionsClient-clone}{\code{AudioTranscriptionsClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AudioTranscriptionsClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-AudioTranscriptionsClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AudioTranscriptionsClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AudioTranscriptionsClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-AudioTranscriptionsClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AudioTranscriptionsClient$create(
  file,
  model = "whisper-1",
  language = NULL,
  prompt = NULL,
  response_format = NULL,
  temperature = NULL,
  timestamp_granularities = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AudioTranscriptionsClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AudioTranscriptionsClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AudioTranscriptionsClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
