% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{AudioTranslationsClient}
\alias{AudioTranslationsClient}
\title{Audio Translations Client}
\description{
Audio Translations Client

Audio Translations Client
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AudioTranslationsClient-new}{\code{AudioTranslationsClient$new()}}
\item \href{#method-AudioTranslationsClient-create}{\code{AudioTranslationsClient$create()}}
\item \href{#method-AudioTranslationsClient-clone}{\code{AudioTranslationsClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AudioTranslationsClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-AudioTranslationsClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AudioTranslationsClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AudioTranslationsClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-AudioTranslationsClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AudioTranslationsClient$create(
  file,
  model = "whisper-1",
  prompt = NULL,
  response_format = NULL,
  temperature = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{Audio file path}

\item{\code{model}}{Model to use (e.g., "whisper-1")}

\item{\code{prompt}}{Optional text to guide translation}

\item{\code{response_format}}{Response format}

\item{\code{temperature}}{Sampling temperature}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Translation response
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AudioTranslationsClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AudioTranslationsClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AudioTranslationsClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
