% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{FilesClient}
\alias{FilesClient}
\title{Files Client}
\description{
Client for the OpenAI Files API. Upload, list, retrieve, and delete files
that can be used with fine-tuning, assistants, batch processing, and more.
Access via \code{client$files}.
}
\section{Methods}{

\describe{
\item{\verb{$create(file, purpose)}}{Upload a file}
\item{\verb{$list(...)}}{List uploaded files, optionally filter by purpose}
\item{\verb{$retrieve(file_id)}}{Get metadata for a specific file}
\item{\verb{$delete(file_id)}}{Delete a file}
\item{\verb{$content(file_id)}}{Download the raw content of a file}
\item{\verb{$wait_for_processing(file_id, ...)}}{Poll until a file is processed}
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilesClient-new}{\code{FilesClient$new()}}
\item \href{#method-FilesClient-create}{\code{FilesClient$create()}}
\item \href{#method-FilesClient-list}{\code{FilesClient$list()}}
\item \href{#method-FilesClient-retrieve}{\code{FilesClient$retrieve()}}
\item \href{#method-FilesClient-delete}{\code{FilesClient$delete()}}
\item \href{#method-FilesClient-content}{\code{FilesClient$content()}}
\item \href{#method-FilesClient-wait_for_processing}{\code{FilesClient$wait_for_processing()}}
\item \href{#method-FilesClient-clone}{\code{FilesClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilesClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilesClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilesClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilesClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-FilesClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilesClient$create(file, purpose)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilesClient-list"></a>}}
\if{latex}{\out{\hypertarget{method-FilesClient-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilesClient$list(purpose = NULL, limit = NULL, after = NULL, order = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilesClient-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-FilesClient-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilesClient$retrieve(file_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilesClient-delete"></a>}}
\if{latex}{\out{\hypertarget{method-FilesClient-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilesClient$delete(file_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilesClient-content"></a>}}
\if{latex}{\out{\hypertarget{method-FilesClient-content}{}}}
\subsection{Method \code{content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilesClient$content(file_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilesClient-wait_for_processing"></a>}}
\if{latex}{\out{\hypertarget{method-FilesClient-wait_for_processing}{}}}
\subsection{Method \code{wait_for_processing()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilesClient$wait_for_processing(file_id, timeout = 300, poll_interval = 5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilesClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilesClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilesClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
