% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_tuning.R
\name{FineTuningCheckpointsClient}
\alias{FineTuningCheckpointsClient}
\title{Fine-tuning Checkpoints Client}
\description{
Retrieves intermediate model checkpoints created during fine-tuning.
A checkpoint is saved after each training epoch. Access via
\code{client$fine_tuning$jobs$checkpoints}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FineTuningCheckpointsClient-new}{\code{FineTuningCheckpointsClient$new()}}
\item \href{#method-FineTuningCheckpointsClient-list}{\code{FineTuningCheckpointsClient$list()}}
\item \href{#method-FineTuningCheckpointsClient-clone}{\code{FineTuningCheckpointsClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineTuningCheckpointsClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-FineTuningCheckpointsClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineTuningCheckpointsClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineTuningCheckpointsClient-list"></a>}}
\if{latex}{\out{\hypertarget{method-FineTuningCheckpointsClient-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineTuningCheckpointsClient$list(
  fine_tuning_job_id,
  after = NULL,
  limit = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineTuningCheckpointsClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FineTuningCheckpointsClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineTuningCheckpointsClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
