% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{\%||\%}
\alias{\%||\%}
\title{Helper function to check NULL}
\usage{
a \%||\% b
}
\arguments{
\item{a}{Value to check}

\item{b}{Default value}
}
\value{
a if not NULL, otherwise b
}
\description{
Helper function to check NULL
}
\keyword{internal}
