% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{handle_response}
\alias{handle_response}
\title{Handle HTTP response and raise appropriate errors}
\usage{
handle_response(resp)
}
\arguments{
\item{resp}{HTTP response object from httr2}
}
\value{
Parsed response content
}
\description{
Handle HTTP response and raise appropriate errors
}
\keyword{internal}
