% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{ImagesClient}
\alias{ImagesClient}
\title{Images Client}
\description{
Client for OpenAI Images API (DALL-E image generation and editing).
Access via \code{client$images}.
}
\section{Methods}{

\describe{
\item{\verb{$create(prompt, ...)}}{Generate a new image from a text description}
\item{\verb{$edit(image, prompt, ...)}}{Edit an existing image using a mask}
\item{\verb{$create_variation(image, ...)}}{Create variations of an existing image}
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImagesClient-new}{\code{ImagesClient$new()}}
\item \href{#method-ImagesClient-create}{\code{ImagesClient$create()}}
\item \href{#method-ImagesClient-edit}{\code{ImagesClient$edit()}}
\item \href{#method-ImagesClient-create_variation}{\code{ImagesClient$create_variation()}}
\item \href{#method-ImagesClient-clone}{\code{ImagesClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagesClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImagesClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagesClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagesClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-ImagesClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagesClient$create(
  prompt,
  model = "dall-e-3",
  n = NULL,
  quality = NULL,
  response_format = NULL,
  size = NULL,
  style = NULL,
  user = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagesClient-edit"></a>}}
\if{latex}{\out{\hypertarget{method-ImagesClient-edit}{}}}
\subsection{Method \code{edit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagesClient$edit(
  image,
  prompt,
  mask = NULL,
  model = "dall-e-2",
  n = NULL,
  response_format = NULL,
  size = NULL,
  user = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagesClient-create_variation"></a>}}
\if{latex}{\out{\hypertarget{method-ImagesClient-create_variation}{}}}
\subsection{Method \code{create_variation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagesClient$create_variation(
  image,
  model = "dall-e-2",
  n = NULL,
  response_format = NULL,
  size = NULL,
  user = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagesClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImagesClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagesClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
