% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{MessagesClient}
\alias{MessagesClient}
\title{Messages Client}
\description{
Manages messages within threads. Messages store the conversation history
between the user and the assistant.
Access via \code{client$threads$messages}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MessagesClient-new}{\code{MessagesClient$new()}}
\item \href{#method-MessagesClient-create}{\code{MessagesClient$create()}}
\item \href{#method-MessagesClient-list}{\code{MessagesClient$list()}}
\item \href{#method-MessagesClient-retrieve}{\code{MessagesClient$retrieve()}}
\item \href{#method-MessagesClient-update}{\code{MessagesClient$update()}}
\item \href{#method-MessagesClient-delete}{\code{MessagesClient$delete()}}
\item \href{#method-MessagesClient-clone}{\code{MessagesClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessagesClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-MessagesClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessagesClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessagesClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-MessagesClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessagesClient$create(
  thread_id,
  role,
  content,
  attachments = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessagesClient-list"></a>}}
\if{latex}{\out{\hypertarget{method-MessagesClient-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessagesClient$list(
  thread_id,
  limit = NULL,
  order = NULL,
  after = NULL,
  before = NULL,
  run_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessagesClient-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-MessagesClient-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessagesClient$retrieve(thread_id, message_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessagesClient-update"></a>}}
\if{latex}{\out{\hypertarget{method-MessagesClient-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessagesClient$update(thread_id, message_id, metadata = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessagesClient-delete"></a>}}
\if{latex}{\out{\hypertarget{method-MessagesClient-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessagesClient$delete(thread_id, message_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessagesClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MessagesClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessagesClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
