% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ModelsClient}
\alias{ModelsClient}
\title{Models Client}
\description{
Client for OpenAI Models API.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelsClient-new}{\code{ModelsClient$new()}}
\item \href{#method-ModelsClient-list}{\code{ModelsClient$list()}}
\item \href{#method-ModelsClient-retrieve}{\code{ModelsClient$retrieve()}}
\item \href{#method-ModelsClient-delete}{\code{ModelsClient$delete()}}
\item \href{#method-ModelsClient-clone}{\code{ModelsClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelsClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelsClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelsClient$new(parent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parent}}{Parent OpenAI client
List available models}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelsClient-list"></a>}}
\if{latex}{\out{\hypertarget{method-ModelsClient-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelsClient$list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Models list response
Retrieve a specific model
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelsClient-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-ModelsClient-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelsClient$retrieve(model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{Model ID (e.g., "gpt-4", "text-embedding-ada-002")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Model details
Delete a fine-tuned model
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelsClient-delete"></a>}}
\if{latex}{\out{\hypertarget{method-ModelsClient-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelsClient$delete(model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{Model ID}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Deletion status
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelsClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelsClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelsClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
