% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderations.R
\name{ModerationsClient}
\alias{ModerationsClient}
\title{Moderations Client}
\description{
Client for the OpenAI Moderations API. Classifies text (and optionally
images) for potentially harmful content according to OpenAI's usage policies.
Access via \code{client$moderations}.
}
\details{
The API returns a set of category flags and confidence scores. It is free
to use and does not count against your token quota.
}
\section{Detected categories}{

\code{hate}, \code{hate/threatening}, \code{harassment}, \code{harassment/threatening},
\code{self-harm}, \code{self-harm/intent}, \code{self-harm/instructions},
\code{sexual}, \code{sexual/minors}, \code{violence}, \code{violence/graphic}.
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModerationsClient-new}{\code{ModerationsClient$new()}}
\item \href{#method-ModerationsClient-create}{\code{ModerationsClient$create()}}
\item \href{#method-ModerationsClient-clone}{\code{ModerationsClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModerationsClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModerationsClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModerationsClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModerationsClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-ModerationsClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModerationsClient$create(input, model = "omni-moderation-latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModerationsClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModerationsClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModerationsClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
