% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{OpenAIAPIError}
\alias{OpenAIAPIError}
\title{OpenAI API Error}
\usage{
OpenAIAPIError(message, status_code = NULL, response = NULL, ...)
}
\arguments{
\item{message}{Error message}

\item{status_code}{HTTP status code}

\item{response}{Raw response object}

\item{...}{Additional arguments}
}
\description{
OpenAI API Error
}
