% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{OpenAIConnectionError}
\alias{OpenAIConnectionError}
\title{OpenAI Connection Error}
\usage{
OpenAIConnectionError(message, ...)
}
\arguments{
\item{message}{Error message}

\item{...}{Additional arguments}
}
\description{
OpenAI Connection Error
}
