% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{OpenAIError}
\alias{OpenAIError}
\title{OpenAI Error Classes}
\usage{
OpenAIError(message, ...)
}
\arguments{
\item{message}{Error message}

\item{...}{Additional arguments passed to \code{\link[rlang:abort]{rlang::abort()}}}
}
\description{
Error classes for OpenAI API errors.
}
\keyword{internal}
