% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{RunStepsClient}
\alias{RunStepsClient}
\title{Run Steps Client}
\description{
Lists and retrieves the individual steps taken during a run.
Each run may consist of multiple steps (e.g. thinking, tool calls,
message creation). Access via \code{client$threads$runs$steps}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RunStepsClient-new}{\code{RunStepsClient$new()}}
\item \href{#method-RunStepsClient-list}{\code{RunStepsClient$list()}}
\item \href{#method-RunStepsClient-retrieve}{\code{RunStepsClient$retrieve()}}
\item \href{#method-RunStepsClient-clone}{\code{RunStepsClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunStepsClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-RunStepsClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunStepsClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunStepsClient-list"></a>}}
\if{latex}{\out{\hypertarget{method-RunStepsClient-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunStepsClient$list(
  thread_id,
  run_id,
  limit = NULL,
  order = NULL,
  after = NULL,
  before = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunStepsClient-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-RunStepsClient-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunStepsClient$retrieve(thread_id, run_id, step_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunStepsClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RunStepsClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunStepsClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
