% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{SpeechClient}
\alias{SpeechClient}
\title{Speech Client (Text-to-Speech)}
\description{
Converts text to spoken audio using OpenAI's TTS models.
Access via \code{client$audio$speech}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SpeechClient-new}{\code{SpeechClient$new()}}
\item \href{#method-SpeechClient-create}{\code{SpeechClient$create()}}
\item \href{#method-SpeechClient-clone}{\code{SpeechClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeechClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-SpeechClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpeechClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeechClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-SpeechClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpeechClient$create(
  input,
  model = "tts-1",
  voice = "alloy",
  response_format = NULL,
  speed = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeechClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SpeechClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpeechClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
