% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stores.R
\name{VectorStoreFileBatchesClient}
\alias{VectorStoreFileBatchesClient}
\title{Vector Store File Batches Client}
\description{
Add multiple files to a vector store in a single batch operation.
More efficient than adding files one-by-one when loading many files.
Access via \code{client$vector_stores$file_batches}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VectorStoreFileBatchesClient-new}{\code{VectorStoreFileBatchesClient$new()}}
\item \href{#method-VectorStoreFileBatchesClient-create}{\code{VectorStoreFileBatchesClient$create()}}
\item \href{#method-VectorStoreFileBatchesClient-retrieve}{\code{VectorStoreFileBatchesClient$retrieve()}}
\item \href{#method-VectorStoreFileBatchesClient-cancel}{\code{VectorStoreFileBatchesClient$cancel()}}
\item \href{#method-VectorStoreFileBatchesClient-list_files}{\code{VectorStoreFileBatchesClient$list_files()}}
\item \href{#method-VectorStoreFileBatchesClient-clone}{\code{VectorStoreFileBatchesClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFileBatchesClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFileBatchesClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFileBatchesClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFileBatchesClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFileBatchesClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFileBatchesClient$create(
  vector_store_id,
  file_ids,
  chunking_strategy = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFileBatchesClient-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFileBatchesClient-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFileBatchesClient$retrieve(vector_store_id, batch_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFileBatchesClient-cancel"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFileBatchesClient-cancel}{}}}
\subsection{Method \code{cancel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFileBatchesClient$cancel(vector_store_id, batch_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFileBatchesClient-list_files"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFileBatchesClient-list_files}{}}}
\subsection{Method \code{list_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFileBatchesClient$list_files(
  vector_store_id,
  batch_id,
  limit = NULL,
  order = NULL,
  after = NULL,
  before = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFileBatchesClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFileBatchesClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFileBatchesClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
