% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stores.R
\name{VectorStoreFilesClient}
\alias{VectorStoreFilesClient}
\title{Vector Store Files Client}
\description{
Manages individual files within a vector store. Adding a file triggers
automatic chunking and embedding.
Access via \code{client$vector_stores$files}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VectorStoreFilesClient-new}{\code{VectorStoreFilesClient$new()}}
\item \href{#method-VectorStoreFilesClient-create}{\code{VectorStoreFilesClient$create()}}
\item \href{#method-VectorStoreFilesClient-list}{\code{VectorStoreFilesClient$list()}}
\item \href{#method-VectorStoreFilesClient-retrieve}{\code{VectorStoreFilesClient$retrieve()}}
\item \href{#method-VectorStoreFilesClient-update}{\code{VectorStoreFilesClient$update()}}
\item \href{#method-VectorStoreFilesClient-delete}{\code{VectorStoreFilesClient$delete()}}
\item \href{#method-VectorStoreFilesClient-content}{\code{VectorStoreFilesClient$content()}}
\item \href{#method-VectorStoreFilesClient-clone}{\code{VectorStoreFilesClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFilesClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFilesClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFilesClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFilesClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFilesClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFilesClient$create(
  vector_store_id,
  file_id,
  chunking_strategy = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFilesClient-list"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFilesClient-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFilesClient$list(
  vector_store_id,
  limit = NULL,
  order = NULL,
  after = NULL,
  before = NULL,
  filter = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFilesClient-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFilesClient-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFilesClient$retrieve(vector_store_id, file_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFilesClient-update"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFilesClient-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFilesClient$update(vector_store_id, file_id, attributes = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFilesClient-delete"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFilesClient-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFilesClient$delete(vector_store_id, file_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFilesClient-content"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFilesClient-content}{}}}
\subsection{Method \code{content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFilesClient$content(vector_store_id, file_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFilesClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFilesClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFilesClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
