% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{AudioClient}
\alias{AudioClient}
\title{Audio Client}
\description{
Client for OpenAI Audio API (transcription, translation, speech).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{transcriptions}}{Audio transcription interface}

\item{\code{translations}}{Audio translation interface}

\item{\code{speech}}{Text-to-speech interface
Initialize audio client}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AudioClient-new}{\code{AudioClient$new()}}
\item \href{#method-AudioClient-clone}{\code{AudioClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AudioClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-AudioClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AudioClient$new(parent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parent}}{Parent OpenAI client}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AudioClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AudioClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AudioClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
