% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_tuning.R
\name{cancel_fine_tuning_job}
\alias{cancel_fine_tuning_job}
\title{Cancel a fine-tuning job (convenience function)}
\usage{
cancel_fine_tuning_job(fine_tuning_job_id)
}
\arguments{
\item{fine_tuning_job_id}{Job ID}
}
\value{
Cancelled job
}
\description{
Cancel a fine-tuning job (convenience function)
}
