% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{create_chat_completion}
\alias{create_chat_completion}
\title{Create a chat completion (convenience function)}
\usage{
create_chat_completion(messages, model = "gpt-3.5-turbo", ...)
}
\arguments{
\item{messages}{List of message objects}

\item{model}{Model to use}

\item{...}{Additional parameters passed to chat$completions$create()}
}
\value{
Chat completion response
}
\description{
Create a chat completion (convenience function)
}
