% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{create_embedding}
\alias{create_embedding}
\title{Create embeddings (convenience function)}
\usage{
create_embedding(input, model = "text-embedding-ada-002", ...)
}
\arguments{
\item{input}{Input text or texts}

\item{model}{Model to use}

\item{...}{Additional parameters}
}
\value{
Embeddings response
}
\description{
Create embeddings (convenience function)
}
