% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{create_image}
\alias{create_image}
\title{Create an image (convenience function)}
\usage{
create_image(prompt, model = "dall-e-3", ...)
}
\arguments{
\item{prompt}{Text description}

\item{model}{Model to use}

\item{...}{Additional parameters}
}
\value{
Images response
}
\description{
Create an image (convenience function)
}
