% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{create_image_edit}
\alias{create_image_edit}
\title{Create an edited image (convenience function)}
\usage{
create_image_edit(image, prompt, mask = NULL, ...)
}
\arguments{
\item{image}{Image file path}

\item{prompt}{Text description}

\item{mask}{Mask file path (optional)}

\item{...}{Additional parameters}
}
\value{
Images response
}
\description{
Create an edited image (convenience function)
}
