% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{create_image_variation}
\alias{create_image_variation}
\title{Create an image variation (convenience function)}
\usage{
create_image_variation(image, model = "dall-e-2", ...)
}
\arguments{
\item{image}{Image file path}

\item{model}{Model to use}

\item{...}{Additional parameters}
}
\value{
Images response
}
\description{
Create an image variation (convenience function)
}
