% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{create_speech}
\alias{create_speech}
\title{Create speech (convenience function)}
\usage{
create_speech(input, model = "tts-1", voice = "alloy", ...)
}
\arguments{
\item{input}{Text to synthesize}

\item{model}{Model to use}

\item{voice}{Voice to use}

\item{...}{Additional parameters}
}
\value{
Raw audio data
}
\description{
Create speech (convenience function)
}
