% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{create_transcription}
\alias{create_transcription}
\title{Create transcription (convenience function)}
\usage{
create_transcription(file, model = "whisper-1", ...)
}
\arguments{
\item{file}{Audio file path}

\item{model}{Model to use}

\item{...}{Additional parameters}
}
\value{
Transcription response
}
\description{
Create transcription (convenience function)
}
