% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{create_translation}
\alias{create_translation}
\title{Create translation (convenience function)}
\usage{
create_translation(file, model = "whisper-1", ...)
}
\arguments{
\item{file}{Audio file path}

\item{model}{Model to use}

\item{...}{Additional parameters}
}
\value{
Translation response
}
\description{
Create translation (convenience function)
}
