% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{delete_model}
\alias{delete_model}
\title{Delete a model (convenience function)}
\usage{
delete_model(model)
}
\arguments{
\item{model}{Model ID}
}
\value{
Deletion status
}
\description{
Delete a model (convenience function)
}
