% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{EmbeddingsClient}
\alias{EmbeddingsClient}
\title{Embeddings Client}
\description{
Client for OpenAI Embeddings API.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EmbeddingsClient-new}{\code{EmbeddingsClient$new()}}
\item \href{#method-EmbeddingsClient-create}{\code{EmbeddingsClient$create()}}
\item \href{#method-EmbeddingsClient-clone}{\code{EmbeddingsClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmbeddingsClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-EmbeddingsClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmbeddingsClient$new(parent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parent}}{Parent OpenAI client
Create embeddings for text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmbeddingsClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-EmbeddingsClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmbeddingsClient$create(
  input,
  model = "text-embedding-ada-002",
  encoding_format = NULL,
  dimensions = NULL,
  user = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{Input text or array of texts}

\item{\code{model}}{Model to use (e.g., "text-embedding-ada-002")}

\item{\code{encoding_format}}{Format for embeddings ("float" or "base64")}

\item{\code{dimensions}}{Number of dimensions for embeddings}

\item{\code{user}}{Unique identifier for end user}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Embeddings response
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmbeddingsClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EmbeddingsClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmbeddingsClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
