% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_tuning.R
\name{FineTuningJobsClient}
\alias{FineTuningJobsClient}
\title{Fine-tuning Jobs Client}
\description{
Fine-tuning Jobs Client

Fine-tuning Jobs Client
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FineTuningJobsClient-new}{\code{FineTuningJobsClient$new()}}
\item \href{#method-FineTuningJobsClient-create}{\code{FineTuningJobsClient$create()}}
\item \href{#method-FineTuningJobsClient-list}{\code{FineTuningJobsClient$list()}}
\item \href{#method-FineTuningJobsClient-retrieve}{\code{FineTuningJobsClient$retrieve()}}
\item \href{#method-FineTuningJobsClient-cancel}{\code{FineTuningJobsClient$cancel()}}
\item \href{#method-FineTuningJobsClient-list_events}{\code{FineTuningJobsClient$list_events()}}
\item \href{#method-FineTuningJobsClient-clone}{\code{FineTuningJobsClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineTuningJobsClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-FineTuningJobsClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineTuningJobsClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineTuningJobsClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-FineTuningJobsClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineTuningJobsClient$create(
  training_file,
  model = "gpt-3.5-turbo",
  hyperparameters = NULL,
  suffix = NULL,
  validation_file = NULL,
  integrations = NULL,
  seed = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{training_file}}{File ID for training data}

\item{\code{model}}{Model to fine-tune (e.g., "gpt-3.5-turbo")}

\item{\code{hyperparameters}}{Hyperparameters for fine-tuning}

\item{\code{suffix}}{Suffix for the fine-tuned model name}

\item{\code{validation_file}}{File ID for validation data}

\item{\code{integrations}}{List of integrations to enable}

\item{\code{seed}}{Random seed}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Fine-tuning job object
List fine-tuning jobs
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineTuningJobsClient-list"></a>}}
\if{latex}{\out{\hypertarget{method-FineTuningJobsClient-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineTuningJobsClient$list(after = NULL, limit = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{after}}{Identifier for the last job from previous request}

\item{\code{limit}}{Number of jobs to retrieve}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of fine-tuning jobs
Retrieve a fine-tuning job
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineTuningJobsClient-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-FineTuningJobsClient-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineTuningJobsClient$retrieve(fine_tuning_job_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fine_tuning_job_id}}{Fine-tuning job ID}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Fine-tuning job details
Cancel a fine-tuning job
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineTuningJobsClient-cancel"></a>}}
\if{latex}{\out{\hypertarget{method-FineTuningJobsClient-cancel}{}}}
\subsection{Method \code{cancel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineTuningJobsClient$cancel(fine_tuning_job_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fine_tuning_job_id}}{Fine-tuning job ID}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Cancelled fine-tuning job
List fine-tuning events
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineTuningJobsClient-list_events"></a>}}
\if{latex}{\out{\hypertarget{method-FineTuningJobsClient-list_events}{}}}
\subsection{Method \code{list_events()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineTuningJobsClient$list_events(
  fine_tuning_job_id,
  after = NULL,
  limit = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fine_tuning_job_id}}{Fine-tuning job ID}

\item{\code{after}}{Identifier for the last event from previous request}

\item{\code{limit}}{Number of events to retrieve}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of fine-tuning events
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineTuningJobsClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FineTuningJobsClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineTuningJobsClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
