% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{ImagesClient}
\alias{ImagesClient}
\title{Images Client}
\description{
Client for OpenAI Images API (DALL-E).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImagesClient-new}{\code{ImagesClient$new()}}
\item \href{#method-ImagesClient-create}{\code{ImagesClient$create()}}
\item \href{#method-ImagesClient-edit}{\code{ImagesClient$edit()}}
\item \href{#method-ImagesClient-create_variation}{\code{ImagesClient$create_variation()}}
\item \href{#method-ImagesClient-clone}{\code{ImagesClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagesClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImagesClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagesClient$new(parent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parent}}{Parent OpenAI client
Create an image}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagesClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-ImagesClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagesClient$create(
  prompt,
  model = "dall-e-3",
  n = NULL,
  quality = NULL,
  response_format = NULL,
  size = NULL,
  style = NULL,
  user = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prompt}}{Text description of the desired image}

\item{\code{model}}{Model to use (e.g., "dall-e-3", "dall-e-2")}

\item{\code{n}}{Number of images to generate}

\item{\code{quality}}{Quality of generated image ("standard" or "hd")}

\item{\code{response_format}}{Response format ("url" or "b64_json")}

\item{\code{size}}{Size of generated image (e.g., "1024x1024")}

\item{\code{style}}{Style of generated image ("vivid" or "natural")}

\item{\code{user}}{Unique identifier for end user}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Images response
Create an edited image
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagesClient-edit"></a>}}
\if{latex}{\out{\hypertarget{method-ImagesClient-edit}{}}}
\subsection{Method \code{edit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagesClient$edit(
  image,
  prompt,
  mask = NULL,
  model = "dall-e-2",
  n = NULL,
  response_format = NULL,
  size = NULL,
  user = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{image}}{Image to edit (file path or raw)}

\item{\code{prompt}}{Text description of desired edit}

\item{\code{mask}}{Mask for the edit (file path or raw)}

\item{\code{model}}{Model to use}

\item{\code{n}}{Number of images to generate}

\item{\code{response_format}}{Response format}

\item{\code{size}}{Size of generated image}

\item{\code{user}}{Unique identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Images response
Create an image variation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagesClient-create_variation"></a>}}
\if{latex}{\out{\hypertarget{method-ImagesClient-create_variation}{}}}
\subsection{Method \code{create_variation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagesClient$create_variation(
  image,
  model = "dall-e-2",
  n = NULL,
  response_format = NULL,
  size = NULL,
  user = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{image}}{Image to vary (file path or raw)}

\item{\code{model}}{Model to use}

\item{\code{n}}{Number of images to generate}

\item{\code{response_format}}{Response format}

\item{\code{size}}{Size of generated image}

\item{\code{user}}{Unique identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Images response
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagesClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImagesClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagesClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
