% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_tuning.R
\name{list_fine_tuning_events}
\alias{list_fine_tuning_events}
\title{List fine-tuning events (convenience function)}
\usage{
list_fine_tuning_events(fine_tuning_job_id, ...)
}
\arguments{
\item{fine_tuning_job_id}{Job ID}

\item{...}{Additional parameters}
}
\value{
List of events
}
\description{
List fine-tuning events (convenience function)
}
