% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_tuning.R
\name{list_fine_tuning_jobs}
\alias{list_fine_tuning_jobs}
\title{List fine-tuning jobs (convenience function)}
\usage{
list_fine_tuning_jobs(...)
}
\arguments{
\item{...}{Additional parameters}
}
\value{
List of fine-tuning jobs
}
\description{
List fine-tuning jobs (convenience function)
}
