% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_tuning.R
\name{retrieve_fine_tuning_job}
\alias{retrieve_fine_tuning_job}
\title{Retrieve a fine-tuning job (convenience function)}
\usage{
retrieve_fine_tuning_job(fine_tuning_job_id)
}
\arguments{
\item{fine_tuning_job_id}{Job ID}
}
\value{
Job details
}
\description{
Retrieve a fine-tuning job (convenience function)
}
