% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{retrieve_model}
\alias{retrieve_model}
\title{Retrieve a model (convenience function)}
\usage{
retrieve_model(model)
}
\arguments{
\item{model}{Model ID}
}
\value{
Model details
}
\description{
Retrieve a model (convenience function)
}
