% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{SpeechClient}
\alias{SpeechClient}
\title{Speech Client}
\description{
Speech Client

Speech Client
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SpeechClient-new}{\code{SpeechClient$new()}}
\item \href{#method-SpeechClient-create}{\code{SpeechClient$create()}}
\item \href{#method-SpeechClient-clone}{\code{SpeechClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeechClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-SpeechClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpeechClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeechClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-SpeechClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpeechClient$create(
  input,
  model = "tts-1",
  voice = "alloy",
  response_format = NULL,
  speed = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{Text to synthesize}

\item{\code{model}}{Model to use (e.g., "tts-1", "tts-1-hd")}

\item{\code{voice}}{Voice to use ("alloy", "echo", "fable", "onyx", "nova", "shimmer")}

\item{\code{response_format}}{Response format ("mp3", "opus", "aac", "flac", "wav", "pcm")}

\item{\code{speed}}{Speech speed (0.25 to 4.0)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Raw audio data (not parsed as JSON)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeechClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SpeechClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpeechClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
